ephemeral: validate provider responses for write-only attributes
